==> 5-turna.ipynb <==
import pandas as pd

df = pd.read_csv("transcription_results_with_metrics_50_tr_md2.csv")

from turkish_lm_tuner import TextPredictor

# import models based on their tasks
sentiment_predictor = TextPredictor(model_name='boun-tabi-LMG/turna_classification_17bintweet_sentiment', task='sentiment')
topic_predictor = TextPredictor(model_name='boun-tabi-LMG/turna_classification_ttc4900', task='categorization')

def analyze_text(text):
    sentiment = sentiment_predictor.predict(text)
    topic = topic_predictor.predict(text)
    return sentiment, topic

# Add new columns for sentiment and topic
df['sentiment'] = ""
df['topic'] = ""

for index, row in df.iterrows():
    transcription = row['transcription']
    sentiment, topic = analyze_text(transcription)
    df.at[index, 'sentiment'] = sentiment
    df.at[index, 'topic'] = topic

df.to_csv("turna1.csv", index=False)
